requirements.txt
requests>=2.31.0
pycryptodome>=3.20.0
tqdm>=4.66.0
